requirements/base.txt
django==2.0.1
djangorestframework==3.7.7
django-environ==0.4.4
django_crontab==0.7.1
django-filter
bootstrap-admin
reportlab
drf-yasg
swagger_spec_validator
flex
python-dateutil
djangodeletes
xlrd
